+++ src/main/kotlin/com/github/gradle/node/yarn/task/YarnTask.kt
@@ -72,6 +72,6 @@ abstract class YarnTask : BaseTask() {
                 NodeExecConfiguration(command, environment.get(), workingDir.asFile.orNull,
                         ignoreExitValue.get(), execOverrides.orNull)
         val yarnExecRunner = objects.newInstance(YarnExecRunner::class.java)
-        result = yarnExecRunner.executeYarnCommand(projectHelper, nodeExtension, nodeExecConfiguration, variantComputer)
+        result = yarnExecRunner.executeYarnCommand(nodeExtension, nodeExecConfiguration, variantComputer)
     }
 }